records observations on own and missing keys only
