+++ test/test.js
@@ -1,3 +1,4 @@
 require("./object-test");
+require("./object-observability-test");
 require("./array-test");
 require("./function-test");
